tests: split CLI integration modules
Split the large mixed tests/test_sbomnix.py module into smaller
concern-specific files for each CLI area, plus shared test path and
helper modules.

Move offline tests into dedicated top-level modules and keep the new
layout focused on test organization only, without application code
changes.
