--- tests/test_sbomnix.py
@@ -1,766 +1,0 @@
-#!/usr/bin/env python3
-# SPDX-FileCopyrightText: 2022-2023 Technology Innovation Institute (TII)
-#
-# SPDX-License-Identifier: Apache-2.0
-
-# pylint: disable=too-few-public-methods
-
-"""Tests for sbomnix"""
-
-import errno
-import json
-import subprocess
-from pathlib import Path
-
-import jsonschema
-import pandas as pd
-import pytest
-import referencing
-import referencing.retrieval
-import requests
-
-from common.utils import df_from_csv_file
-from provenance import main as provenance_main
-from vulnxscan.whitelist import df_apply_whitelist, load_whitelist
-
-MYDIR = Path(__file__).resolve().parent
-
-# These two tools are executed as scripts from this file,
-# but don't contain any tests themselves
-COMPARE_DEPS = MYDIR / "compare_deps.py"
-COMPARE_SBOMS = MYDIR / "compare_sboms.py"
-
-REPOROOT = MYDIR / ".."
-SRCDIR = REPOROOT / "src"
-
-# The different entrypoints of the application. Currently we invoke them with a
-# new python interpreter, as it's not in a shape yet to import a function we
-# pass arguments to.
-SBOMNIX = SRCDIR / "sbomnix" / "main.py"
-NIXGRAPH = SRCDIR / "nixgraph" / "main.py"
-NIXMETA = SRCDIR / "nixmeta" / "main.py"
-PROVENANCE = SRCDIR / "provenance" / "main.py"
-NIX_OUTDATED = SRCDIR / "nixupdate" / "nix_outdated.py"
-VULNXSCAN = SRCDIR / "vulnxscan" / "vulnxscan_cli.py"
-REPOLOGY_CLI = SRCDIR / "repology" / "repology_cli.py"
-REPOLOGY_CVE = SRCDIR / "repology" / "repology_cve.py"
-
-
-def test_sbomnix_help(_run_python_script):
-    """Test sbomnix command line argument: '-h'"""
-    _run_python_script([SBOMNIX, "-h"])
-
-
-def test_sbomnix_type_runtime(_run_python_script, test_nix_result, test_work_dir):
-    """Test sbomnix generates valid CycloneDX json with runtime dependencies"""
-    out_path_cdx = test_work_dir / "sbom_cdx_test.json"
-    out_path_spdx = test_work_dir / "sbom_spdx_test.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx.as_posix(),
-            "--spdx",
-            out_path_spdx.as_posix(),
-        ]
-    )
-    assert out_path_cdx.exists()
-    assert out_path_spdx.exists()
-    cdx_schema_path = MYDIR / "resources" / "cdx_bom-1.4.schema.json"
-    assert cdx_schema_path.exists()
-    validate_json(out_path_cdx.as_posix(), cdx_schema_path)
-    spdx_schema_path = MYDIR / "resources" / "spdx_bom-2.3.schema.json"
-    assert spdx_schema_path.exists()
-    validate_json(out_path_spdx.as_posix(), spdx_schema_path)
-
-
-@pytest.mark.slow
-def test_sbomnix_type_buildtime(_run_python_script, test_nix_result, test_work_dir):
-    """Test sbomnix generates valid CycloneDX json with buildtime dependencies"""
-    out_path_cdx = test_work_dir / "sbom_cdx_test.json"
-    out_path_spdx = test_work_dir / "sbom_spdx_test.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx.as_posix(),
-            "--spdx",
-            out_path_spdx.as_posix(),
-            "--buildtime",
-        ]
-    )
-    assert out_path_cdx.exists()
-    assert out_path_spdx.exists()
-    cdx_schema_path = MYDIR / "resources" / "cdx_bom-1.4.schema.json"
-    assert cdx_schema_path.exists()
-    validate_json(out_path_cdx.as_posix(), cdx_schema_path)
-    spdx_schema_path = MYDIR / "resources" / "spdx_bom-2.3.schema.json"
-    assert spdx_schema_path.exists()
-    validate_json(out_path_spdx.as_posix(), spdx_schema_path)
-
-
-def test_sbomnix_depth(_run_python_script, test_nix_result, test_work_dir):
-    """Test sbomnix '--depth' option"""
-    out_path_csv_1 = test_work_dir / "sbom_csv_test_1.csv"
-    out_path_csv_2 = test_work_dir / "sbom_csv_test_2.csv"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--buildtime",
-            "--csv",
-            out_path_csv_1.as_posix(),
-            "--depth=2",
-        ]
-    )
-    assert out_path_csv_1.exists()
-    df_out_1 = pd.read_csv(out_path_csv_1)
-    assert not df_out_1.empty
-
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--buildtime",
-            "--csv",
-            out_path_csv_2.as_posix(),
-            "--depth=1",
-        ]
-    )
-    assert out_path_csv_2.exists()
-    df_out_2 = pd.read_csv(out_path_csv_2)
-    assert not df_out_2.empty
-    # Check the dataframes are not equal
-    df_diff = df_difference(df_out_1, df_out_2)
-    assert not df_diff.empty, df_to_string(df_diff)
-    df_right_only = df_diff[df_diff["_merge"] == "right_only"]
-    # Check df_out_1 contains rows that are not in df_out_2
-    assert df_right_only.empty, df_to_string(df_diff)
-
-
-################################################################################
-
-
-def test_nixgraph_help(_run_python_script):
-    """Test nixgraph command line argument: '-h'"""
-    _run_python_script([NIXGRAPH, "-h"])
-
-
-def test_nixgraph_png(_run_python_script, test_nix_result, test_work_dir):
-    """Test nixgraph with png output generates valid png image"""
-    png_out = test_work_dir / "graph.png"
-    _run_python_script([NIXGRAPH, test_nix_result, "--out", png_out, "--depth", "3"])
-    assert Path(png_out).exists()
-
-
-def test_nixgraph_csv(_run_python_script, test_nix_result, test_work_dir):
-    """Test nixgraph with csv output generates valid csv"""
-    csv_out = test_work_dir / "graph.csv"
-    _run_python_script([NIXGRAPH, test_nix_result, "--out", csv_out, "--depth", "3"])
-    assert Path(csv_out).exists()
-    # Check the output is valid csv file
-    df_out = pd.read_csv(csv_out)
-    assert not df_out.empty
-
-
-def test_nixgraph_csv_buildtime(_run_python_script, test_nix_result, test_work_dir):
-    """Test nixgraph with buildtime csv output generates valid csv"""
-    csv_out = test_work_dir / "graph_buildtime.csv"
-    _run_python_script([NIXGRAPH, test_nix_result, "--out", csv_out, "--buildtime"])
-    assert Path(csv_out).exists()
-    # Check the output is valid csv file
-    df_out = pd.read_csv(csv_out)
-    assert not df_out.empty
-
-
-def test_nixgraph_csv_graph_inverse(_run_python_script, test_nix_result, test_work_dir):
-    """Test nixgraph with '--inverse' argument"""
-    csv_out = test_work_dir / "graph.csv"
-    _run_python_script(
-        [
-            NIXGRAPH,
-            test_nix_result,
-            "--out",
-            csv_out,
-            "--depth=100",
-        ]
-    )
-    assert Path(csv_out).exists()
-    df_out = pd.read_csv(csv_out)
-    assert not df_out.empty
-
-    csv_out_inv = test_work_dir / "graph_inverse.csv"
-    _run_python_script(
-        [
-            NIXGRAPH,
-            test_nix_result,
-            "--out",
-            csv_out_inv,
-            "--depth=100",
-            "--inverse=.*",
-        ]
-    )
-    assert Path(csv_out_inv).exists()
-    df_out_inv = pd.read_csv(csv_out_inv)
-    assert not df_out_inv.empty
-
-    # When 'depth' covers the entire graph, the output from
-    # the two above commands should be the same, except for column
-    # 'graph_depth': below, we remove that column from both outputs and
-    # compare the two dataframes
-    df_out = df_out.drop("graph_depth", axis=1)
-    df_out = df_out.sort_values(by=["src_path"])
-    df_out_inv = df_out_inv.drop("graph_depth", axis=1)
-    df_out_inv = df_out_inv.sort_values(by=["src_path"])
-    df_diff = df_difference(df_out, df_out_inv)
-    assert df_diff.empty, df_to_string(df_diff)
-
-
-################################################################################
-
-
-def test_compare_deps_runtime(_run_python_script, test_nix_result, test_work_dir):
-    """Compare nixgraph vs sbom runtime dependencies"""
-    graph_csv_out = test_work_dir / "graph.csv"
-    _run_python_script(
-        [
-            NIXGRAPH,
-            test_nix_result,
-            "--out",
-            graph_csv_out,
-            "--depth=100",
-        ]
-    )
-    assert Path(graph_csv_out).exists()
-
-    out_path_cdx = test_work_dir / "sbom_cdx_test.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx.as_posix(),
-        ]
-    )
-    assert out_path_cdx.exists()
-
-    _run_python_script(
-        [
-            COMPARE_DEPS,
-            "--sbom",
-            out_path_cdx,
-            "--graph",
-            graph_csv_out,
-        ]
-    )
-
-
-@pytest.mark.slow
-@pytest.mark.skip_in_ci
-def test_compare_deps_buildtime(_run_python_script, test_nix_result, test_work_dir):
-    """Compare nixgraph vs sbom buildtime dependencies"""
-    graph_csv_out = test_work_dir / "graph.csv"
-    _run_python_script(
-        [
-            NIXGRAPH,
-            test_nix_result,
-            "--out",
-            graph_csv_out,
-            "--depth=100",
-            "--buildtime",
-        ]
-    )
-    assert Path(graph_csv_out).exists()
-
-    out_path_cdx = test_work_dir / "sbom_cdx_test.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx.as_posix(),
-            "--buildtime",
-        ]
-    )
-    assert out_path_cdx.exists()
-
-    _run_python_script(
-        [
-            COMPARE_DEPS,
-            "--sbom",
-            out_path_cdx,
-            "--graph",
-            graph_csv_out,
-        ]
-    )
-
-
-@pytest.mark.slow
-def test_compare_subsequent_cdx_sboms(
-    _run_python_script, test_nix_result, test_work_dir
-):
-    """Compare two sbomnix runs with same target produce the same cdx sbom"""
-    out_path_cdx_1 = test_work_dir / "sbom_cdx_test_1.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx_1.as_posix(),
-            "--buildtime",
-        ]
-    )
-    assert out_path_cdx_1.exists()
-
-    out_path_cdx_2 = test_work_dir / "sbom_cdx_test_2.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx_2.as_posix(),
-            "--buildtime",
-        ]
-    )
-    assert out_path_cdx_2.exists()
-
-    _run_python_script(
-        [
-            COMPARE_SBOMS,
-            out_path_cdx_1,
-            out_path_cdx_2,
-        ]
-    )
-
-
-@pytest.mark.slow
-def test_compare_subsequent_spdx_sboms(
-    _run_python_script, test_nix_result, test_work_dir
-):
-    """Compare two sbomnix runs with same target produce the same spdx sbom"""
-    out_path_spdx_1 = test_work_dir / "sbom_spdx_test_1.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--spdx",
-            out_path_spdx_1.as_posix(),
-            "--buildtime",
-        ]
-    )
-    assert out_path_spdx_1.exists()
-
-    out_path_spdx_2 = test_work_dir / "sbom_spdx_test_2.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--spdx",
-            out_path_spdx_2.as_posix(),
-            "--buildtime",
-        ]
-    )
-    assert out_path_spdx_2.exists()
-
-    _run_python_script(
-        [
-            COMPARE_SBOMS,
-            out_path_spdx_1,
-            out_path_spdx_2,
-        ]
-    )
-
-
-@pytest.mark.slow
-def test_compare_spdx_and_cdx_sboms(_run_python_script, test_nix_result, test_work_dir):
-    """Compare spdx and cdx sboms from the same sbomnix invocation"""
-    out_path_spdx = test_work_dir / "sbom_spdx_test.json"
-    out_path_cdx = test_work_dir / "sbom_cdx_test.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx.as_posix(),
-            "--spdx",
-            out_path_spdx.as_posix(),
-            "--buildtime",
-        ]
-    )
-    assert out_path_cdx.exists()
-    assert out_path_spdx.exists()
-
-    _run_python_script(
-        [
-            COMPARE_SBOMS,
-            out_path_cdx,
-            out_path_spdx,
-        ]
-    )
-
-
-################################################################################
-
-
-def test_vulnxscan_help(_run_python_script):
-    """Test vulnxscan command line argument: '-h'"""
-    _run_python_script([VULNXSCAN, "--help"])
-
-
-@pytest.mark.skip_in_ci
-def test_vulnxscan_scan_nix_result(_run_python_script, test_nix_result, test_work_dir):
-    """Test vulnxscan scan with TEST_NIX_RESULT as input"""
-    out_path_vulns = test_work_dir / "vulnxscan_test.csv"
-    _run_python_script(
-        [
-            VULNXSCAN,
-            test_nix_result.as_posix(),
-            "--out",
-            out_path_vulns.as_posix(),
-        ]
-    )
-
-
-@pytest.mark.skip_in_ci
-def test_vulnxscan_scan_sbom(_run_python_script, test_nix_result, test_work_dir):
-    """Test vulnxscan scan with SBOM as input"""
-    out_path_cdx = test_work_dir / "sbom_cdx_test.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx,
-        ]
-    )
-    assert out_path_cdx.exists()
-
-    out_path_vulns = test_work_dir / "vulnxscan_test.csv"
-    _run_python_script(
-        [
-            VULNXSCAN,
-            "--sbom",
-            out_path_cdx.as_posix(),
-            "--out",
-            out_path_vulns.as_posix(),
-        ]
-    )
-
-
-def test_vulnxscan_triage(_run_python_script, test_nix_result, test_work_dir):
-    """Test vulnxscan scan with --triage"""
-    out_path_vulns = test_work_dir / "vulnxscan_test.csv"
-    _run_python_script(
-        [
-            VULNXSCAN,
-            "--triage",
-            "--out",
-            out_path_vulns.as_posix(),
-            test_nix_result.as_posix(),
-        ]
-    )
-
-
-@pytest.mark.skip_in_ci
-def test_vulnxscan_triage_whitelist(_run_python_script, test_nix_result, test_work_dir):
-    """Test vulnxscan scan with --triage and --whitelist"""
-    out_path_vulns = test_work_dir / "vulnxscan_test.csv"
-    whitelist_csv = MYDIR / "resources" / "whitelist_all.csv"
-    assert whitelist_csv.exists()
-    ret = _run_python_script(
-        [
-            VULNXSCAN,
-            "--triage",
-            "--whitelist",
-            whitelist_csv.as_posix(),
-            "--out",
-            out_path_vulns.as_posix(),
-            test_nix_result.as_posix(),
-        ],
-        capture_output=True,
-        text=True,
-    )
-    # Console output should not include any vulnerabilities
-    # when whitelisting all vulnerabilities
-    assert "Potential vulnerabilities impacting version_local" not in ret.stderr
-
-
-################################################################################
-
-
-def test_repology_cli_help(_run_python_script):
-    """
-    Test repology_cli command line argument: '-h'
-    """
-    _run_python_script([REPOLOGY_CLI, "-h"])
-
-
-@pytest.mark.slow
-def test_repology_cli_sbom(
-    _run_python_script,
-    _run_python_script_retry_on_repology_network_error,
-    test_nix_result,
-    test_work_dir,
-):
-    """Test repology_cli with SBOM as input"""
-    out_path_cdx = test_work_dir / "sbom_cdx_test.json"
-    _run_python_script(
-        [
-            SBOMNIX,
-            test_nix_result,
-            "--cdx",
-            out_path_cdx,
-        ]
-    )
-    assert out_path_cdx.exists()
-
-    out_path_repology = test_work_dir / "repology.csv"
-    _run_python_script_retry_on_repology_network_error(
-        [
-            REPOLOGY_CLI,
-            "--sbom_cdx",
-            out_path_cdx.as_posix(),
-            "--repository",
-            "nix_unstable",
-            "--out",
-            out_path_repology.as_posix(),
-        ]
-    )
-    assert out_path_repology.exists()
-
-
-################################################################################
-
-
-def test_nix_outdated_help(_run_python_script):
-    """
-    Test nix_outdated command line argument: '-h'
-    """
-    _run_python_script([NIX_OUTDATED, "-h"])
-
-
-@pytest.mark.slow
-def test_nix_outdated_result(
-    _run_python_script_retry_on_repology_network_error, test_nix_result, test_work_dir
-):
-    """Test nix_outdated with TEST_NIX_RESULT as input"""
-    out_path_nix_outdated = test_work_dir / "nix_outdated.csv"
-    _run_python_script_retry_on_repology_network_error(
-        [
-            NIX_OUTDATED,
-            "--out",
-            out_path_nix_outdated.as_posix(),
-            test_nix_result,
-        ]
-    )
-    assert out_path_nix_outdated.exists()
-
-
-################################################################################
-
-
-def test_nixmeta_help(_run_python_script):
-    """
-    Test nixmeta command line argument: '-h'
-    """
-    _run_python_script([NIXMETA, "-h"])
-
-
-@pytest.mark.slow
-def test_nixmeta_sbomnix_flakeref(_run_python_script, test_work_dir):
-    """Test nixmeta with sbomnix flakeref"""
-    out_path = test_work_dir / "nixmeta.csv"
-    _run_python_script(
-        [
-            NIXMETA,
-            "--out",
-            out_path.as_posix(),
-            "--flakeref",
-            REPOROOT,
-        ]
-    )
-    assert out_path.exists()
-    # Quick sanity checks for the output data
-    df_meta = df_from_csv_file(out_path)
-    assert df_meta is not None
-    entries = df_meta.shape[0]
-    assert entries > 50000
-
-
-################################################################################
-
-
-def test_whitelist():
-    """Test applying whitelist to vulnerability csv file"""
-    # Load test files
-    whitelist_csv = MYDIR / "resources" / "whitelist.csv"
-    assert whitelist_csv.exists()
-    vulns_csv = MYDIR / "resources" / "vulns.csv"
-    assert vulns_csv.exists()
-    df_whitelist = load_whitelist(whitelist_csv)
-    assert df_whitelist is not None
-    df_vulns = df_from_csv_file(vulns_csv)
-    assert df_vulns is not None
-    print("df_vulns:")
-    print(df_vulns.info())
-    print(df_vulns)
-    print("whitelist:")
-    print(df_whitelist)
-    # Copy df_vulns, including only the columns we want to include in the diff
-    df_vuln_id_copy = df_vulns.copy()[["vuln_id", "package"]]
-    print("df_vuln_id_copy:")
-    print(df_vuln_id_copy)
-    # Apply whitelist, this changes df_vuln_id_copy in-place
-    # by adding columns "whitelist" and "whitelist_comment"
-    df_apply_whitelist(df_whitelist, df_vuln_id_copy)
-    print("df_vuln_id_copy after whitelist apply")
-    print(df_vuln_id_copy.info())
-    print(df_vuln_id_copy)
-    # After applying whitelist, the resulting dataframe should match df_vulns
-    df_diff = df_difference(df_vulns.astype(str), df_vuln_id_copy.astype(str))
-    print("diff")
-    print(df_diff)
-    assert df_diff.empty, df_to_string(df_diff)
-
-
-################################################################################
-
-
-def test_provenance_help(_run_python_script):
-    """Test provenance command line argument: '-h'"""
-    _run_python_script([PROVENANCE, "-h"])
-
-
-def test_provenance_schema(_run_python_script, test_nix_result, test_work_dir):
-    """Test provenance generates valid schema"""
-    out_path = test_work_dir / "provenance_test.json"
-    _run_python_script(
-        [
-            PROVENANCE,
-            test_nix_result,
-            "--out",
-            out_path.as_posix(),
-        ]
-    )
-    assert out_path.exists()
-    schema_path = MYDIR / "resources" / "provenance-1.0.schema.json"
-    assert schema_path.exists()
-    validate_json(out_path.as_posix(), schema_path)
-
-
-def test_provenance_schema_recursive(
-    _run_python_script, test_nix_result, test_work_dir
-):
-    """Test provenance generates valid schema with recursive option"""
-    out_path = test_work_dir / "recursive_provenance_test.json"
-    _run_python_script(
-        [
-            PROVENANCE,
-            test_nix_result,
-            "--recursive",
-            "--out",
-            out_path.as_posix(),
-        ]
-    )
-    assert out_path.exists()
-    schema_path = MYDIR / "resources" / "provenance-1.0.schema.json"
-    assert schema_path.exists()
-    validate_json(out_path.as_posix(), schema_path)
-
-
-def test_provenance_hash_query_batches_on_e2big(monkeypatch):
-    """Test provenance splits oversized nix-store hash queries and preserves order"""
-
-    references = [f"/nix/store/hash-{idx}" for idx in range(5)]
-    calls = []
-
-    def fake_exec_cmd(cmd, **_kwargs):
-        if cmd[:3] == ["nix-store", "--query", "--hash"]:
-            batch = cmd[3:]
-            calls.append(batch)
-            if len(batch) > 2:
-                raise OSError(errno.E2BIG, "Argument list too long")
-            hashes = "\n".join(
-                f"sha256:hash-{path.rsplit('-', 1)[-1]}" for path in batch
-            )
-            return subprocess.CompletedProcess(cmd, 0, stdout=f"{hashes}\n", stderr="")
-        raise AssertionError(f"unexpected command: {cmd}")
-
-    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
-
-    hashes = provenance_main.query_store_hashes(references)
-
-    assert hashes == [f"sha256:hash-{idx}" for idx in range(5)]
-    assert calls == [
-        references,
-        references[:2],
-        references[2:],
-        references[2:3],
-        references[3:],
-    ]
-
-
-################################################################################
-
-
-class JSONSchemaRetrieve:
-    """Cached retriever that can be used with jsonschema.validate"""
-
-    @staticmethod
-    @referencing.retrieval.to_cached_resource()
-    def _retrieve_via_requests(uri):
-        print(f"retrieving schema: {uri}")
-        return requests.get(uri, timeout=10).text
-
-    def __call__(self, uri):
-        return self._retrieve_via_requests(uri)
-
-
-def validate_json(file_path, schema_path):
-    """Validate json file matches schema"""
-    with (
-        open(file_path, encoding="utf-8") as json_file,
-        open(schema_path, encoding="utf-8") as schema_file,
-    ):
-        json_obj = json.load(json_file)
-        schema_obj = json.load(schema_file)
-        reg = referencing.Registry(retrieve=JSONSchemaRetrieve())
-        jsonschema.validate(json_obj, schema_obj, registry=reg)
-
-
-def df_to_string(df):
-    """Convert dataframe to string"""
-    return (
-        "\n"
-        + df.to_string(max_rows=None, max_cols=None, index=False, justify="left")
-        + "\n"
-    )
-
-
-def df_difference(df_left, df_right):
-    """Return dataframe that represents diff of two dataframes"""
-    df_right = df_right.astype(df_left.dtypes.to_dict())
-    df = df_left.merge(
-        df_right,
-        how="outer",
-        indicator=True,
-    )
-    # Keep only the rows that differ (that are not in both)
-    df = df[df["_merge"] != "both"]
-    # Re-order columns: last column ('_merge') becomes first
-    cols = df.columns.tolist()
-    cols = cols[-1:] + cols[:-1]
-    df = df[cols]
-    return df
-
-
-################################################################################
-
-
-if __name__ == "__main__":
-    pytest.main([__file__])
-
-
-################################################################################
